django/tutorial_project/src/requirements/base.txt
Django>=3.2
psycopg2-binary==2.9.6
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2